refactor: move data transformations out of viewmodel breaking change! fixing asap
